requirements.txt
aiofiles==23.2.1
aiohttp==3.9.3
aiohttp-sse-client==0.2.1
aiosignal==1.3.1
annotated-types==0.6.0
async-timeout==4.0.3
attrs==23.2.0
bitarray==2.9.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
discord.py==2.3.2
frozenlist==1.4.1
idna==3.6
magic-filter==1.0.12
multidict==6.0.5
pycparser==2.21
pycryptodomex==3.20.0
pydantic==2.5.3
pydantic_core==2.14.6
PyNaCl==1.5.0
pypng==0.20220715.0
python-dotenv==1.0.1
pytonconnect==0.3.0
pytoniq==0.1.32
pytoniq-core==0.1.27
qrcode==7.4.2
redis==5.0.1
requests==2.31.0
typing_extensions==4.9.0
urllib3==2.2.1
x25519==0.0.2
yarl==1.9.4
